requirements.txt
streamlit
opencv-python
opencv-contrib-python
numpy
pillow
torch
torchvision
